Make volume transitions sound smoother.
Perform linear interpolation between each 10ms-spaced volume control
point, so as to make the changes in volume less jarring to the listener.

GstControlSource holds onto all the control-points given to it until
they are explicitly removed.  To prevent the amount of storage required
by these from growing large, we remove points older than 1 second.  This
strategy might remove points which are still needed if the amount of
buffering delay on the main audio is larger than the buffering delay on
the control channel (i.e. if the difference is more than 1 second).